+++ src/index.js
@@ -1,0 +1,1 @@
+export { rebind } from './rebind'
